+++ tests/testthat/setup.R
@@ -1,13 +1,8 @@
 has_tern_key <- function() {
-  tryCatch(
-    {
-      key <- keyring::key_get(
-        service = "TERN_API_KEY",
-        keyring = "nert"
-      )
-
-      nzchar(key)
-    },
-    error = function(e) FALSE
+  key <- tryCatch(
+    get_key(),
+    error = function(e) ""
   )
+
+  nzchar(key)
 }